Moved some enum defines for kerncaps to kc.h
