+++ include/nn/hac/define/kc.h
@@ -23,6 +23,169 @@ namespace hac
 			KC_MISC_FLAGS = 16,
 			KC_STUB = 32
 		};
+
+		enum MiscFlags
+		{
+			MISCFLAG_ENABLE_DEBUG,
+			MISCFLAG_FORCE_DEBUG,
+			MISCFLAG_UNK02,
+			MISCFLAG_UNK03,
+			MISCFLAG_UNK04,
+			MISCFLAG_UNK05,
+			MISCFLAG_UNK06,
+			MISCFLAG_UNK07,
+			MISCFLAG_UNK08,
+			MISCFLAG_UNK09,
+			MISCFLAG_UNK10,
+			MISCFLAG_UNK11,
+			MISCFLAG_UNK12,
+			MISCFLAG_UNK13,
+			MISCFLAG_UNK14,
+			MISCFLAG_NUM
+		};
+
+		enum MemoryPerm
+		{
+			MEM_RW,
+			MEM_RO
+		};
+
+		enum MappingType
+		{
+			MAP_IO,
+			MAP_STATIC
+		};
+
+		enum SystemCall
+		{
+			SVC_SET_HEAP_SIZE = 0x01,
+			SVC_SET_MEMORY_PERMISSION,
+			SVC_SET_MEMORY_ATTRIBUTE,
+			SVC_MAP_MEMORY,
+			SVC_UNMAP_MEMORY,
+			SVC_QUERY_MEMORY,
+			SVC_EXIT_PROCESS,
+			SVC_CREATE_THREAD,
+			SVC_START_THREAD,
+			SVC_EXIT_THREAD,
+			SVC_SLEEP_THREAD,
+			SVC_GET_THREAD_PRIORITY,
+			SVC_SET_THREAD_PRIORITY,
+			SVC_GET_THREAD_CORE_MASK,
+			SVC_SET_THREAD_CORE_MASK,
+			SVC_GET_CURRENT_PROCESSOR_NUMBER = 0x10,
+			SVC_SIGNAL_EVENT,
+			SVC_CLEAR_EVENT,
+			SVC_MAP_SHARED_MEMORY,
+			SVC_UNMAP_SHARED_MEMORY,
+			SVC_CREATE_TRANSFER_MEMORY,
+			SVC_CLOSE_HANDLE,
+			SVC_RESET_SIGNAL,
+			SVC_WAIT_SYNCHRONIZATION,
+			SVC_CANCEL_SYNCHRONIZATION,
+			SVC_ARBIRATE_LOCK,
+			SVC_ARBIRATE_UNLOCK,
+			SVC_WAIT_PROCESS_WIDE_KEY_ATOMIC,
+			SVC_SIGNAL_PROCESS_WIDE_KEY,
+			SVC_GET_SYSTEM_TICK,
+			SVC_CONNECT_TO_NAMED_PORT,
+			SVC_SEND_SYNC_REQUEST_LIGHT = 0x20,
+			SVC_SEND_SYNC_REQUEST,
+			SVC_SEND_SYNC_REQUEST_WITH_USER_BUFFER,
+			SVC_SEND_ASYNC_REQUEST_WITH_USER_BUFFER,
+			SVC_GET_PROCESS_ID,
+			SVC_GET_THREAD_ID,
+			SVC_BREAK,
+			SVC_OUTPUT_DEBUG_STRING,
+			SVC_RETURN_FROM_EXCEPTION,
+			SVC_GET_INFO,
+			SVC_FLUSH_ENTIRE_DATA_CACHE,
+			SVC_FLUSH_DATA_CACHE,
+			SVC_MAP_PHYSICAL_MEMORY,
+			SVC_UNMAP_PHYSICAL_MEMORY,
+			SVC_GET_FUTURE_THREAD_INFO,
+			SVC_GET_LAST_THREAD_INFO,
+			SVC_GET_RESOURCE_LIMIT_LIMIT_VALUE = 0x30,
+			SVC_GET_RESOURCE_LIMIT_CURRENT_VALUE,
+			SVC_SET_THREAD_ACTIVITY,
+			SVC_GET_THREAD_CONTEXT3,
+			SVC_WAIT_FOR_ADDRESS,
+			SVC_SIGNAL_TO_ADDRESS,
+			SVC_SYNCHRONIZE_PREEMPTION_STATE,
+			SVC_UNK_0x37,
+			SVC_UNK_0x38,
+			SVC_UNK_0x39,
+			SVC_UNK_0x3A,
+			SVC_UNK_0x3B,
+			SVC_DUMP_INFO,
+			SVC_DUMP_INFO_NEW,
+			SVC_UNK_0x3E,
+			SVC_UNK_0x3F,
+			SVC_CREATE_SESSION = 0x40,
+			SVC_ACCEPT_SESSION,
+			SVC_REPLY_AND_RECEIVE_LIGHT,
+			SVC_REPLY_AND_RECEIVE,
+			SVC_REPLY_AND_RECEIVE_WITH_USER_BUFFER,
+			SVC_CREATE_EVENT,
+			SVC_UNK_0x46,
+			SVC_UNK_0x47,
+			SVC_MAP_PHYSICAL_MEMORY_UNSAFE,
+			SVC_UNMAP_PHYSICAL_MEMORY_UNSAFE,
+			SVC_SET_UNSAFE_LIMIT,
+			SVC_CREATE_CODE_MEMORY,
+			SVC_CONTROL_CODE_MEMORY,
+			SVC_SLEEP_SYSTEM,
+			SVC_READ_WRITE_REGISTER,
+			SVC_SET_PROCESS_ACTIVITY,
+			SVC_CREATE_SHARED_MEMORY = 0x50,
+			SVC_MAP_TRANSFER_MEMORY,
+			SVC_UNMAP_TRANSFER_MEMORY,
+			SVC_CREATE_INTERRUPT_EVENT,
+			SVC_QUERY_PHYSICAL_ADDRESS,
+			SVC_QUERY_IO_MAPPING,
+			SVC_CREATE_DEVICE_ADDRESS_SPACE,
+			SVC_ATTACH_DEVICE_ADDRESS_SPACE,
+			SVC_DETACH_DEVICE_ADDRESS_SPACE,
+			SVC_MAP_DEVICE_ADDRESS_SPACE_BY_FORCE,
+			SVC_MAP_DEVICE_ADDRESS_SPACE_ALIGNED,
+			SVC_MAP_DEVICE_ADDRESS_SPACE,
+			SVC_UNMAP_DEVICE_ADDRESS_SPACE,
+			SVC_INVALIDATE_PROCESS_DATA_CACHE,
+			SVC_STORE_PROCESS_DATA_CACHE,
+			SVC_FLUSH_PROCESS_DATA_CACHE,
+			SVC_DEBUG_ACTIVE_PROCESS = 0x60,
+			SVC_BREAK_DEBUG_PROCESS,
+			SVC_TERMINATE_DEBUG_PROCESS,
+			SVC_GET_DEBUG_EVENT,
+			SVC_CONTINUE_DEBUG_EVENT,
+			SVC_GET_PROCESS_LIST,
+			SVC_GET_THREAD_LIST,
+			SVC_GET_DEBUG_THREAD_CONTEXT,
+			SVC_SET_DEBUG_THREAD_CONTEXT,
+			SVC_QUERY_DEBUG_PROCESS_MEMORY,
+			SVC_READ_DEBUG_PROCESS_MEMORY,
+			SVC_WRITE_DEBUG_PROCESS_MEMORY,
+			SVC_SET_HARDWARE_BREAK_POINT,
+			SVC_GET_DEBUG_THREAD_PARAM,
+			SVC_UNK_0x6E,
+			SVC_GET_SYSTEM_INFO,
+			SVC_CREATE_PORT = 0x70,
+			SVC_MANAGE_NAMED_PORT,
+			SVC_CONNECT_TO_PORT,
+			SVC_SET_PROCESS_MEMORY_PERMISSION,
+			SVC_MAP_PROCESS_MEMORY,
+			SVC_UNMAP_PROCESS_MEMORY,
+			SVC_QUERY_PROCESS_MEMORY,
+			SVC_MAP_PROCESS_CODE_MEMORY,
+			SVC_UNMAP_PROCESS_CODE_MEMORY,
+			SVC_CREATE_PROCESS,
+			SVC_START_PROCESS,
+			SVC_TERMINATE_PROCESS,
+			SVC_GET_PROCESS_INFO,
+			SVC_CREATE_RESOURCE_LIMIT,
+			SVC_SET_RESOURCE_LIMIT_LIMIT_VALUE,
+			SVC_CALL_SECURE_MONITOR,
+		};
 	}
 }
 }
